separated game and render logic, fixes #188
